move pywincffi.kernel32.io.GetStdHandle -> pywincffi.kernel32.handle.GetStdHandle
