+++ pywincffi/kernel32/io.py
@@ -332,34 +332,3 @@ def CloseHandle(hObject):
     code = library.CloseHandle(hObject)
     error_check("CloseHandle", code=code, expected=Enums.NON_ZERO)
 
-
-def GetStdHandle(nStdHandle):
-    """
-    Retrieves a handle to the specified standard
-    device (standard input, standard output, or standard error).
-
-    .. seealso::
-
-        https://msdn.microsoft.com/en-us/library/ms683231
-
-    :param int nStdHandle:
-        The standard device to retrieve
-
-    :rtype: handle
-    :return:
-        Returns a handle to the standard device retrieved.
-    """
-    _, library = dist.load()
-    input_check("nStdHandle", nStdHandle,
-                allowed_values=(library.STD_INPUT_HANDLE,
-                                library.STD_OUTPUT_HANDLE,
-                                library.STD_ERROR_HANDLE))
-
-    handle = library.GetStdHandle(nStdHandle)
-
-    if handle == INVALID_HANDLE_VALUE:  # pragma: no cover
-        raise WindowsAPIError(
-            "GetStdHandle", "Invalid Handle", INVALID_HANDLE_VALUE,
-            "not %s" % INVALID_HANDLE_VALUE)
-
-    return handle
