+++ tests/test_kernel32/test_handle.py
@@ -1,0 +1,26 @@
+from pywincffi.core import dist
+from pywincffi.dev.testutil import TestCase
+from pywincffi.kernel32.handle import GetStdHandle
+
+
+class TestGetStdHandle(TestCase):
+    def test_stdin_handle(self):
+        _, library = dist.load()
+        self.assertEqual(
+            GetStdHandle(library.STD_INPUT_HANDLE),
+            library.GetStdHandle(library.STD_INPUT_HANDLE)
+        )
+
+    def test_stdout_handle(self):
+        _, library = dist.load()
+        self.assertEqual(
+            GetStdHandle(library.STD_OUTPUT_HANDLE),
+            library.GetStdHandle(library.STD_OUTPUT_HANDLE)
+        )
+
+    def test_stderr_handle(self):
+        _, library = dist.load()
+        self.assertEqual(
+            GetStdHandle(library.STD_ERROR_HANDLE),
+            library.GetStdHandle(library.STD_ERROR_HANDLE)
+        )
